TrafficFlowForecast/requirements.txt
absl-py==0.10.0
apache-beam==2.24.0
APScheduler==3.6.3
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
autopep8==1.5.4
avro-python3==1.9.2.1
backcall==0.2.0
bleach==3.2.1
cachetools==3.1.1
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
crcmod==1.7
cx-Oracle==8.1.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docker==4.3.1
docopt==0.6.2
elasticsearch==7.8.0
entrypoints==0.3
fastavro==0.21.24
fasteners==0.15
Flask==1.1.2
Flask-APScheduler==1.11.0
future==0.18.2
gast==0.2.2
google-api-core==1.22.4
google-api-python-client==1.12.3
google-apitools==0.5.31
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-build==2.0.0
google-cloud-core==1.4.3
google-cloud-datastore==1.7.4
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.0.2
google-cloud-spanner==1.19.0
google-cloud-videointelligence==1.16.0
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==0.4.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.33.1
grpcio-gcp==0.2.2
h5py==2.10.0
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets==7.5.1
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-tuner==1.0.1
kubernetes==11.0.0
lazy-object-proxy==1.4.3
libcst==0.3.13
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
ml-metadata==0.24.0
mock==2.0.0
monotonic==1.5
mypy-extensions==0.4.3
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numexpr==2.7.1
numpy==1.19.2
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pbr==5.5.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.7
proto-plus==1.11.0
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pyarrow==0.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
Pygments==2.7.1
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.21.3
scipy==1.5.3
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
tables==3.6.1
tabulate==0.8.7
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-data-validation==0.24.1
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-metadata==0.24.0
tensorflow-model-analysis==0.24.3
tensorflow-serving-api==1.15.0
tensorflow-transform==0.24.1
termcolor==1.1.0
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
tfx==0.24.1
tfx-bsl==0.24.1
toml==0.10.1
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.4.0
